Fix schema test
